+++ RELEASES.md
@@ -88,6 +88,7 @@ Upcoming
 - Add version to "Upload Configuration" and add derivative for name and version from file name
 - Bugfix security log file for rest service "Test a PipeLine"
 - Use correct configuration in "Show Environment variables"
+- Add javaListenerSessionKey to IbisLocalSender